Stage 1/5: Your program will receive command line arguments in this order:
IP address
port
message for sending
The algorithm is the following:

Create a new socket.
Connect to a host and a port using the socket.
Send a message from the third command line argument to the host using the socket.
Receive the server’s response.
Print the server’s response.
Close the socket.